fix connect hardwear transfer crash,fix bugly report bug
