+++ android/blelibrary/src/main/java/cn/com/heaton/blelibrary/ble/utils/BleUtils.java
@@ -1,12 +1,10 @@
 package cn.com.heaton.blelibrary.ble.utils;
-
 import android.app.ActivityManager;
 import android.content.Context;
 import android.os.Build;
 import android.provider.Settings;
 import android.text.TextUtils;
 
-import java.util.HashMap;
 import java.util.List;
 
 /**
@@ -20,12 +18,14 @@ public class BleUtils {
                 .getSystemService(Context.ACTIVITY_SERVICE);
         List<ActivityManager.RunningAppProcessInfo> appProcesses = activityManager
                 .getRunningAppProcesses();
-        for (ActivityManager.RunningAppProcessInfo appProcess : appProcesses) {
-            if (appProcess.processName.equals(context.getPackageName())) {
-                if (appProcess.importance != ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND) {
-                    return true;
-                } else {
-                    return false;
+        if (appProcesses != null && appProcesses.size() > 0) {
+            for (ActivityManager.RunningAppProcessInfo appProcess : appProcesses) {
+                if (appProcess.processName.equals(context.getPackageName())) {
+                    if (appProcess.importance != ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND) {
+                        return true;
+                    } else {
+                        return false;
+                    }
                 }
             }
         }
